refactor(api): expose the generation store and MCP registry as effect-free contracts + factories
