+++ .changeset/485-api-dev-runtime-protocol.md
@@ -2,4 +2,4 @@
 'agent-bundle': patch
 ---
 
-Export the `dev.runtime.provider` protocol from `agent-bundle/api`, next to `DevRuntimeProvider`: the start-context and session types (`DevRuntimeStartContext`, `DevRuntimeSession`, `DevRuntimePreparedProject`, `DevRuntimePreparedMcpServer`, `DevRuntimePreparedMcpApp`, `DevRuntimeEventInput`, `DevRuntimeClientSurfaceEndpoint`, `DevRuntimeMcpRegistry`, `DevRuntimeMcpSession`, `DevRuntimeMcpSessionView`, `DevRuntimeMcpSessionCloseObservation`, `DevRuntimeMcpSessionExecuteOptions`, `DevRuntimeMcpRegistryListener`, `DevRuntimeMcpRegistryMessage`, `DevRuntimeMcpRegistrySubscription`), the wire protocol (`DevRuntimeInspectionEnvelope`, `DevRuntimeMcpServerDescriptor`, `DevRuntimeRun`, `DevRuntimeStatus`, `DevRuntimeSurface`, `DevRuntimeDescriptor`, `DevRuntimeDiagnostic`, `DevRuntimeDiagnosticPhase`, `DevRuntimeFixture`, `DevRuntimeTraceSpan`, `DevRuntimeTreeNode`, `DevRuntimeAsset`, `DevRuntimeAssetRequest`, `DevRuntimeInvocationRequest`, `DevRuntimeReplayRequest`, `DevRuntimeStateIdentity`, `DevRuntimeStateResetRequest`, `DevRuntimeMcpAppRunBinding`, `DevRuntimeMcpConnectionState`, `DevRuntimeMcpInvalidatedBinding`, `DevRuntimeMcpOperationRequest`, `DevRuntimeMcpOperationResult`, `DevRuntimeMcpRegistryReconcileInput`, `DevRuntimeMcpRegistryReconcileResult`, `DevRuntimeMcpRegistryReplayGap`, `DevRuntimeMcpRegistrySnapshot`, `DevRuntimeMcpSessionBinding`, `DevRuntimeMcpSessionControlRequest`, `DevRuntimeMcpSessionRequest`, `DevRuntimeMcpSessionSnapshot`, `RuntimeVector`), the errors a provider throws (`DevRuntimeUnavailableError` AB8201, `DevRuntimeGenerationConflictError` AB8204), and the `RuntimeGenerationStore` and `RuntimeMcpRegistry` classes a session drives with their option, candidate, lease, and error types (`RuntimeGenerationStoreError`, `RuntimeGenerationStoreCloseError`, `RuntimeMcpRegistryError`, `RuntimeMcpRegistryCloseError`). A provider can now be written out of tree against `agent-bundle/api` alone; `examples/rsc-agent-runtime/src` no longer reaches into the package source. Fixes #485. (#528)
+Export the `dev.runtime.provider` protocol from `agent-bundle/api`, next to `DevRuntimeProvider`: the start-context and session types (`DevRuntimeStartContext`, `DevRuntimeSession`, `DevRuntimePreparedProject`, `DevRuntimePreparedMcpServer`, `DevRuntimePreparedMcpApp`, `DevRuntimeEventInput`, `DevRuntimeClientSurfaceEndpoint`, `DevRuntimeMcpRegistry`, `DevRuntimeMcpSession`, `DevRuntimeMcpSessionView`, `DevRuntimeMcpSessionCloseObservation`, `DevRuntimeMcpSessionExecuteOptions`, `DevRuntimeMcpRegistryListener`, `DevRuntimeMcpRegistryMessage`, `DevRuntimeMcpRegistrySubscription`), the wire protocol (`DevRuntimeInspectionEnvelope`, `DevRuntimeMcpServerDescriptor`, `DevRuntimeRun`, `DevRuntimeStatus`, `DevRuntimeSurface`, `DevRuntimeDescriptor`, `DevRuntimeDiagnostic`, `DevRuntimeDiagnosticPhase`, `DevRuntimeFixture`, `DevRuntimeTraceSpan`, `DevRuntimeTreeNode`, `DevRuntimeAsset`, `DevRuntimeAssetRequest`, `DevRuntimeInvocationRequest`, `DevRuntimeReplayRequest`, `DevRuntimeStateIdentity`, `DevRuntimeStateResetRequest`, `DevRuntimeMcpAppRunBinding`, `DevRuntimeMcpConnectionState`, `DevRuntimeMcpInvalidatedBinding`, `DevRuntimeMcpOperationRequest`, `DevRuntimeMcpOperationResult`, `DevRuntimeMcpRegistryReconcileInput`, `DevRuntimeMcpRegistryReconcileResult`, `DevRuntimeMcpRegistryReplayGap`, `DevRuntimeMcpRegistrySnapshot`, `DevRuntimeMcpSessionBinding`, `DevRuntimeMcpSessionControlRequest`, `DevRuntimeMcpSessionRequest`, `DevRuntimeMcpSessionSnapshot`, `RuntimeVector`), the errors a provider throws (`DevRuntimeUnavailableError` AB8201, `DevRuntimeGenerationConflictError` AB8204), and the generation store and MCP registry a session drives as effect-free contracts with constructors — `DevRuntimeGenerationStore` / `createRuntimeGenerationStore`, `DevRuntimeProviderMcpRegistry` / `createRuntimeMcpRegistry` — plus their option, candidate, lease, manifest, and error-code types (`RuntimeGenerationStoreOptions`, `RuntimeGenerationCandidate`, `RuntimeGenerationLease`, `RuntimeGenerationManifest`, `RuntimeGenerationStoreErrorCode`, `RuntimeMcpRegistryOptions`, `RuntimeMcpConnector`, `RuntimeMcpExecutionContext`, `RuntimeMcpRegistryErrorCode`, …). A provider can now be written out of tree against `agent-bundle/api` alone; `examples/rsc-agent-runtime/src` no longer reaches into the package source. Fixes #485. (#528)
