+++ packages/agent-bundle/src/dev/runtime-generation-store.ts
@@ -14,93 +14,46 @@ import { isAbsolute, join, relative, resolve, sep } from 'node:path';
 import { digest, stableJson } from '../core/digest.ts';
 import { freezeJsonValue, type JsonObject, type JsonValue } from './types.ts';
 import { YieldableFrameworkError } from '../effect/errors.ts';
+import type {
+  DevRuntimeGenerationStore,
+  RuntimeGeneration,
+  RuntimeGenerationAsset,
+  RuntimeGenerationCandidate,
+  RuntimeGenerationCloseFailure,
+  RuntimeGenerationLease,
+  RuntimeGenerationManifest,
+  RuntimeGenerationManifestInput,
+  RuntimeGenerationMetadataCodec,
+  RuntimeGenerationPrepareOptions,
+  RuntimeGenerationPreparedActivation,
+  RuntimeGenerationStoreErrorCode,
+  RuntimeGenerationStoreOptions,
+  RuntimeGenerationValidator,
+} from './runtime-store-contracts.ts';
+
+// The contracts live in `runtime-store-contracts.ts` (effect-free, exported
+// from `agent-bundle/api`); re-exported here for the internal callers.
+export type {
+  DevRuntimeGenerationStore,
+  RuntimeGeneration,
+  RuntimeGenerationActivationGuard,
+  RuntimeGenerationAsset,
+  RuntimeGenerationCandidate,
+  RuntimeGenerationCloseFailure,
+  RuntimeGenerationLease,
+  RuntimeGenerationManifest,
+  RuntimeGenerationManifestInput,
+  RuntimeGenerationMetadataCodec,
+  RuntimeGenerationPrepareOptions,
+  RuntimeGenerationPreparedActivation,
+  RuntimeGenerationStoreErrorCode,
+  RuntimeGenerationStoreOptions,
+  RuntimeGenerationValidationInput,
+  RuntimeGenerationValidator,
+} from './runtime-store-contracts.ts';
 
 const manifestFileName = 'generation.manifest.json';
 const defaultRetainInactive = 5;
-
-export interface RuntimeGenerationAsset {
-  readonly bytes: number;
-  readonly path: string;
-  readonly sha256: string;
-}
-
-export interface RuntimeGenerationMetadataCodec<TMetadata> {
-  decode(value: JsonValue): TMetadata;
-  encode(value: TMetadata): JsonValue;
-}
-
-export interface RuntimeGenerationManifestInput<TMetadata = unknown> {
-  readonly assets: readonly RuntimeGenerationAsset[];
-  readonly metadata: TMetadata;
-}
-
-export interface RuntimeGenerationManifest<TMetadata = unknown> {
-  readonly assets: readonly RuntimeGenerationAsset[];
-  readonly createdAt: string;
-  readonly id: string;
-  readonly manifestDigest: string;
-  readonly metadata: TMetadata;
-  readonly schemaVersion: 1;
-  readonly sourceRevision: string;
-}
-
-export interface RuntimeGenerationValidationInput<TMetadata> {
-  readonly assets: readonly RuntimeGenerationAsset[];
-  readonly metadata: TMetadata;
-  readonly root: string;
-}
-
-export type RuntimeGenerationValidator<TMetadata> = (
-  input: RuntimeGenerationValidationInput<TMetadata>,
-) => Promise<TMetadata> | TMetadata;
-
-export interface RuntimeGenerationActivationGuard<TMetadata> {
-  wait(manifest: RuntimeGenerationManifest<TMetadata>): Promise<void>;
-  check(manifest: RuntimeGenerationManifest<TMetadata>): boolean;
-}
-
-export interface RuntimeGenerationPrepareOptions<TMetadata> {
-  readonly guard?: RuntimeGenerationActivationGuard<TMetadata>;
-}
-
-export interface RuntimeGenerationCandidate {
-  readonly id: string;
-  readonly root: string;
-  readonly sequence: number;
-  readonly sourceRevision: string;
-}
-
-export interface RuntimeGenerationPreparedActivation<TMetadata = unknown> {
-  readonly generation: RuntimeGeneration<TMetadata>;
-  readonly sequence: number;
-}
-
-export interface RuntimeGeneration<TMetadata = unknown> {
-  readonly id: string;
-  readonly manifest: RuntimeGenerationManifest<TMetadata>;
-  readonly root: string;
-  readonly sourceRevision: string;
-}
-
-export interface RuntimeGenerationLease<TMetadata = unknown> {
-  readonly generation: RuntimeGeneration<TMetadata>;
-  release(): Promise<void>;
-}
-
-export interface RuntimeGenerationStoreOptions<TMetadata> {
-  readonly metadataCodec: RuntimeGenerationMetadataCodec<TMetadata>;
-  readonly now?: () => Date;
-  /** Test seam for cleanup failures; production callers use recursive `rm`. */
-  readonly remove?: (path: string) => Promise<void>;
-  readonly retainInactive?: number;
-  readonly storageRoot: string;
-  readonly validateMetadata: RuntimeGenerationValidator<TMetadata>;
-}
-
-export interface RuntimeGenerationCloseFailure {
-  readonly error: unknown;
-  readonly path: string;
-}
 
 export class RuntimeGenerationStoreCloseError extends YieldableFrameworkError {
   readonly failures: readonly RuntimeGenerationCloseFailure[];
@@ -111,13 +64,6 @@ export class RuntimeGenerationStoreCloseError extends YieldableFrameworkError {
     this.failures = Object.freeze([...failures]);
   }
 }
-
-export type RuntimeGenerationStoreErrorCode =
-  | 'RUNTIME_GENERATION_CLOSED'
-  | 'RUNTIME_GENERATION_CONFLICT'
-  | 'RUNTIME_GENERATION_INVALID'
-  | 'RUNTIME_GENERATION_NOT_FOUND'
-  | 'RUNTIME_GENERATION_SUPERSEDED';
 
 export class RuntimeGenerationStoreError extends YieldableFrameworkError {
   readonly code: RuntimeGenerationStoreErrorCode;
@@ -285,7 +231,7 @@ const fsync = async (path: string): Promise<void> => {
  * Provider-neutral immutable generations. The store accepts only JSON through
  * a provider-supplied codec and never interprets provider metadata keys.
  */
-export class RuntimeGenerationStore<TMetadata = unknown> {
+export class RuntimeGenerationStore<TMetadata = unknown> implements DevRuntimeGenerationStore<TMetadata> {
   readonly #candidates = new WeakMap<RuntimeGenerationCandidate, CandidateRecord>();
   readonly #committed = new Map<string, CommittedRecord<TMetadata>>();
   readonly #generationsRoot: string;
