+++ website/docs/zh/reference/configuration.mdx
@@ -159,7 +159,7 @@ export default defineConfig({
 | `dev.agentApi` | 从 `agent-bundle dev` 暴露那个经过认证、仅 loopback 的 Agent API。`--agent-api` / `--no-agent-api` 标志会覆盖它。 |
 | `dev.contracts.fixtures` | **设置了 `dev.contracts` 时必填。** 项目相对路径的模块，其默认导出把路由 id 映射到契约夹具。声明 `dev.contracts` 会让 `agent-bundle dev` 从直接采用每个 epoch，改为以开发契约矩阵门控面向宿主的采用：检查失败的 epoch 仍会发布到 Workbench playground，但活跃的宿主连接与开发安装会保留最后一个通过的 epoch（`AB7211`）。块本身格式错误、夹具模块逃出项目根、无法加载或导出了错误的形状，则是 `AB7210`。 |
 | `dev.contracts.server` | 矩阵要检查的 MCP 服务器。仅当项目恰好编译一个服务器时可省略。 |
-| `dev.runtime.provider` | 开发期运行时 provider 模块：它导出 `createDevRuntimeProvider`（[`CreateDevRuntimeProvider`](../api/types/api.CreateDevRuntimeProvider.md)），返回一个 [`DevRuntimeProvider`](../api/interfaces/api.DevRuntimeProvider.md)，其 `start` 接收 [`DevRuntimeStartContext`](../api/interfaces/api.DevRuntimeStartContext.md) 并返回 [`DevRuntimeSession`](../api/interfaces/api.DevRuntimeSession.md)。provider 需要实现的整个协议都从 `agent-bundle/api` 发布——检视信封（[`DevRuntimeInspectionEnvelope`](../api/interfaces/api.DevRuntimeInspectionEnvelope.md)）、MCP 服务器描述符（[`DevRuntimeMcpServerDescriptor`](../api/interfaces/api.DevRuntimeMcpServerDescriptor.md)）、它要抛出的错误（[`DevRuntimeUnavailableError`](../api/classes/api.DevRuntimeUnavailableError.md)、[`DevRuntimeGenerationConflictError`](../api/classes/api.DevRuntimeGenerationConflictError.md)），以及会话所驱动的 [`RuntimeGenerationStore`](../api/classes/api.RuntimeGenerationStore.md) 与 [`RuntimeMcpRegistry`](../api/classes/api.RuntimeMcpRegistry.md)。 |
+| `dev.runtime.provider` | 开发期运行时 provider 模块：它导出 `createDevRuntimeProvider`（[`CreateDevRuntimeProvider`](../api/types/api.CreateDevRuntimeProvider.md)），返回一个 [`DevRuntimeProvider`](../api/interfaces/api.DevRuntimeProvider.md)，其 `start` 接收 [`DevRuntimeStartContext`](../api/interfaces/api.DevRuntimeStartContext.md) 并返回 [`DevRuntimeSession`](../api/interfaces/api.DevRuntimeSession.md)。provider 需要实现的整个协议都从 `agent-bundle/api` 发布——检视信封（[`DevRuntimeInspectionEnvelope`](../api/interfaces/api.DevRuntimeInspectionEnvelope.md)）、MCP 服务器描述符（[`DevRuntimeMcpServerDescriptor`](../api/interfaces/api.DevRuntimeMcpServerDescriptor.md)）、它要抛出的错误（[`DevRuntimeUnavailableError`](../api/classes/api.DevRuntimeUnavailableError.md)、[`DevRuntimeGenerationConflictError`](../api/classes/api.DevRuntimeGenerationConflictError.md)），以及会话所驱动的 generation store 与 MCP registry——以契约（[`DevRuntimeGenerationStore`](../api/interfaces/api.DevRuntimeGenerationStore.md)、[`DevRuntimeProviderMcpRegistry`](../api/interfaces/api.DevRuntimeProviderMcpRegistry.md)）及其构造函数（[`createRuntimeGenerationStore`](../api/functions/api.createRuntimeGenerationStore.md)、[`createRuntimeMcpRegistry`](../api/functions/api.createRuntimeMcpRegistry.md)）的形式提供。 |
 
 ## tools
